requirements.txt
requests
ddddocr
pycrypto